+++ packages/sizuku/src/cli/index.test.ts
@@ -2,7 +2,7 @@ import fs from "node:fs";
 import os from "node:os";
 import path from "node:path";
 import { afterEach, beforeEach, describe, expect, it } from "vite-plus/test";
-import { detectOutputType, parseFlags, resolveSchemaLibrary, stripImports } from "./index.js";
+import { stripImports } from "../utils/index.js";
 
 const HELP_TEXT = `💧 sizuku - Drizzle ORM schema tools
 
@@ -76,89 +76,6 @@ export const account = pgTable('account', {
   isActive: boolean('is_active').notNull(),
   createdAt: timestamp('created_at').defaultNow(),
 })`;
-
-// ============================================================================
-// detectOutputType
-// ============================================================================
-
-describe("detectOutputType", () => {
-  it("returns dbml for .dbml extension", () => {
-    expect(detectOutputType("schema.dbml")).toBe("dbml");
-  });
-
-  it("returns png for .png extension", () => {
-    expect(detectOutputType("er-diagram.png")).toBe("png");
-  });
-
-  it("returns mermaid for .md extension", () => {
-    expect(detectOutputType("ER.md")).toBe("mermaid");
-  });
-
-  it("returns typescript for .ts extension", () => {
-    expect(detectOutputType("output.ts")).toBe("typescript");
-  });
-
-  it("returns null for unsupported extension .json", () => {
-    expect(detectOutputType("output.json")).toBe(null);
-  });
-
-  it("returns null for unsupported extension .sql", () => {
-    expect(detectOutputType("output.sql")).toBe(null);
-  });
-
-  it("returns null for unsupported extension .svg", () => {
-    expect(detectOutputType("diagram.svg")).toBe(null);
-  });
-
-  it("returns null for empty string", () => {
-    expect(detectOutputType("")).toBe(null);
-  });
-
-  it("returns null for filename without extension", () => {
-    expect(detectOutputType("README")).toBe(null);
-  });
-
-  it("returns null for dot-only filename", () => {
-    expect(detectOutputType(".")).toBe(null);
-  });
-
-  it("handles nested path with .dbml", () => {
-    expect(detectOutputType("docs/db/schema.dbml")).toBe("dbml");
-  });
-
-  it("handles nested path with .png", () => {
-    expect(detectOutputType("docs/er/diagram.png")).toBe("png");
-  });
-
-  it("handles nested path with .md", () => {
-    expect(detectOutputType("docs/mermaid/ER.md")).toBe("mermaid");
-  });
-
-  it("handles deeply nested output path", () => {
-    expect(detectOutputType("a/b/c/d/e/schema.dbml")).toBe("dbml");
-  });
-
-  it("returns null for .dbml.bak (compound extension)", () => {
-    expect(detectOutputType("schema.dbml.bak")).toBe(null);
-  });
-
-  it("returns null for .md in directory name but .txt file", () => {
-    expect(detectOutputType("docs.md/schema.txt")).toBe(null);
-  });
-
-  it("returns null for .png in directory name but .json file", () => {
-    expect(detectOutputType("images.png/data.json")).toBe(null);
-  });
-
-  it("handles absolute path with .dbml", () => {
-    expect(detectOutputType("/home/user/project/schema.dbml")).toBe("dbml");
-  });
-
-  it("handles absolute path with .md", () => {
-    expect(detectOutputType("/tmp/output/ER.md")).toBe("mermaid");
-  });
-});
-
 // ============================================================================
 // stripImports
 // ============================================================================
@@ -251,147 +168,6 @@ export const post = mysqlTable('post', {
     expect(result[4]).toBe("export const post = mysqlTable('post', {");
   });
 });
-
-// ============================================================================
-// parseFlags
-// ============================================================================
-
-describe("parseFlags", () => {
-  it("returns all defaults when no flags", () => {
-    expect(parseFlags(["schema.ts", "-o", "out.ts"])).toStrictEqual({
-      zod: false,
-      valibot: false,
-      arktype: false,
-      effect: false,
-      zodVersion: undefined,
-      exportTypes: true,
-      withComment: true,
-      withRelation: true,
-    });
-  });
-
-  it("parses --zod flag", () => {
-    const flags = parseFlags(["schema.ts", "-o", "out.ts", "--zod"]);
-    expect(flags.zod).toBe(true);
-    expect(flags.valibot).toBe(false);
-  });
-
-  it("parses --valibot flag", () => {
-    const flags = parseFlags(["schema.ts", "-o", "out.ts", "--valibot"]);
-    expect(flags.valibot).toBe(true);
-    expect(flags.zod).toBe(false);
-  });
-
-  it("parses --arktype flag", () => {
-    const flags = parseFlags(["schema.ts", "-o", "out.ts", "--arktype"]);
-    expect(flags.arktype).toBe(true);
-  });
-
-  it("parses --effect flag", () => {
-    const flags = parseFlags(["schema.ts", "-o", "out.ts", "--effect"]);
-    expect(flags.effect).toBe(true);
-  });
-
-  it("parses --no-export-types", () => {
-    const flags = parseFlags(["--zod", "--no-export-types"]);
-    expect(flags.exportTypes).toBe(false);
-  });
-
-  it("parses --no-with-comment", () => {
-    const flags = parseFlags(["--zod", "--no-with-comment"]);
-    expect(flags.withComment).toBe(false);
-  });
-
-  it("parses --no-with-relation", () => {
-    const flags = parseFlags(["--zod", "--no-with-relation"]);
-    expect(flags.withRelation).toBe(false);
-  });
-
-  it("parses --zod-version=mini (equals syntax)", () => {
-    const flags = parseFlags(["--zod", "--zod-version=mini"]);
-    expect(flags.zodVersion).toBe("mini");
-  });
-
-  it("parses --zod-version v4 (space syntax)", () => {
-    const flags = parseFlags(["--zod", "--zod-version", "v4"]);
-    expect(flags.zodVersion).toBe("v4");
-  });
-
-  it("parses --zod-version=@hono/zod-openapi", () => {
-    const flags = parseFlags(["--zod", "--zod-version=@hono/zod-openapi"]);
-    expect(flags.zodVersion).toBe("@hono/zod-openapi");
-  });
-
-  it("returns undefined zodVersion for invalid value", () => {
-    const flags = parseFlags(["--zod", "--zod-version=invalid"]);
-    expect(flags.zodVersion).toBe(undefined);
-  });
-
-  it("combines multiple flags", () => {
-    const flags = parseFlags([
-      "--valibot",
-      "--no-export-types",
-      "--no-with-comment",
-      "--no-with-relation",
-    ]);
-    expect(flags).toStrictEqual({
-      zod: false,
-      valibot: true,
-      arktype: false,
-      effect: false,
-      zodVersion: undefined,
-      exportTypes: false,
-      withComment: false,
-      withRelation: false,
-    });
-  });
-});
-
-// ============================================================================
-// resolveSchemaLibrary
-// ============================================================================
-
-describe("resolveSchemaLibrary", () => {
-  it("returns zod when --zod flag is set", () => {
-    expect(resolveSchemaLibrary(parseFlags(["--zod"]))).toStrictEqual({
-      name: "zod",
-      label: "Zod",
-    });
-  });
-
-  it("returns valibot when --valibot flag is set", () => {
-    expect(resolveSchemaLibrary(parseFlags(["--valibot"]))).toStrictEqual({
-      name: "valibot",
-      label: "Valibot",
-    });
-  });
-
-  it("returns arktype when --arktype flag is set", () => {
-    expect(resolveSchemaLibrary(parseFlags(["--arktype"]))).toStrictEqual({
-      name: "arktype",
-      label: "ArkType",
-    });
-  });
-
-  it("returns effect when --effect flag is set", () => {
-    expect(resolveSchemaLibrary(parseFlags(["--effect"]))).toStrictEqual({
-      name: "effect",
-      label: "Effect",
-    });
-  });
-
-  it("returns null when no library flag is set", () => {
-    expect(resolveSchemaLibrary(parseFlags([]))).toBe(null);
-  });
-
-  it("returns zod when both --zod and --valibot are set (first wins)", () => {
-    expect(resolveSchemaLibrary(parseFlags(["--zod", "--valibot"]))).toStrictEqual({
-      name: "zod",
-      label: "Zod",
-    });
-  });
-});
-
 // ============================================================================
 // sizuku() - argv parsing integration (via process.argv mock)
 // ============================================================================
@@ -664,232 +440,6 @@ describe("sizuku() direct schema via process.argv", () => {
 // ============================================================================
 // E2E: DBML generation
 // ============================================================================
-
-describe("E2E: DBML generation", () => {
-  let tmpdir: string;
-
-  beforeEach(() => {
-    tmpdir = fs.mkdtempSync(path.join(os.tmpdir(), "sizuku-e2e-dbml-"));
-  });
-
-  afterEach(() => {
-    fs.rmSync(tmpdir, { recursive: true, force: true });
-  });
-
-  it("generates DBML with correct table and column definitions for EC schema", async () => {
-    fs.writeFileSync(path.join(tmpdir, "schema.ts"), ecSchemaContent, "utf-8");
-    const output = path.join(tmpdir, "output", "schema.dbml");
-
-    const code = stripImports(ecSchemaContent);
-    const { sizukuDbml } = await import("../generator/dbml/index.js");
-    const result = await sizukuDbml(code, output);
-
-    expect(result.ok).toBe(true);
-    expect(fs.existsSync(output)).toBe(true);
-
-    const content = fs.readFileSync(output, "utf-8");
-
-    // Tables exist
-    expect(content.includes("Table user {")).toBe(true);
-    expect(content.includes("Table order {")).toBe(true);
-    expect(content.includes("Table orderItem {")).toBe(true);
-
-    // User columns
-    expect(content.includes("id varchar [pk]")).toBe(true);
-    expect(content.includes("name varchar")).toBe(true);
-    expect(content.includes("email varchar")).toBe(true);
-
-    // Order columns
-    expect(content.includes("userId varchar")).toBe(true);
-    expect(content.includes("total decimal")).toBe(true);
-    expect(content.includes("createdAt timestamp")).toBe(true);
-
-    // OrderItem columns
-    expect(content.includes("orderId int")).toBe(true);
-    expect(content.includes("productName varchar")).toBe(true);
-    expect(content.includes("quantity int")).toBe(true);
-    expect(content.includes("price decimal")).toBe(true);
-
-    // Foreign key refs
-    expect(content.includes("Ref order_userId_user_id_fk: order.userId > user.id")).toBe(true);
-    expect(
-      content.includes("Ref orderItem_orderId_order_id_fk: orderItem.orderId > order.id"),
-    ).toBe(true);
-  });
-
-  it("generates DBML with field descriptions as notes for auth schema", async () => {
-    fs.writeFileSync(path.join(tmpdir, "schema.ts"), pgAuthSchema, "utf-8");
-    const output = path.join(tmpdir, "output", "auth.dbml");
-
-    const code = stripImports(pgAuthSchema);
-    const { sizukuDbml } = await import("../generator/dbml/index.js");
-    const result = await sizukuDbml(code, output);
-
-    expect(result.ok).toBe(true);
-    const content = fs.readFileSync(output, "utf-8");
-
-    expect(content.includes("Table account {")).toBe(true);
-    expect(content.includes("note: 'Primary key'")).toBe(true);
-    expect(content.includes("note: 'User email address'")).toBe(true);
-  });
-
-  it("generates DBML for single-table schema without refs", async () => {
-    fs.writeFileSync(path.join(tmpdir, "schema.ts"), singleTableSchema, "utf-8");
-    const output = path.join(tmpdir, "output", "config.dbml");
-
-    const code = stripImports(singleTableSchema);
-    const { sizukuDbml } = await import("../generator/dbml/index.js");
-    const result = await sizukuDbml(code, output);
-
-    expect(result.ok).toBe(true);
-    const content = fs.readFileSync(output, "utf-8");
-
-    expect(content.includes("Table config {")).toBe(true);
-    expect(content.includes("key text [pk]")).toBe(true);
-    expect(content.includes("value text")).toBe(true);
-    // No Ref lines
-    expect(content.includes("Ref ")).toBe(false);
-  });
-});
-
-// ============================================================================
-// E2E: Mermaid ER generation
-// ============================================================================
-
-describe("E2E: Mermaid ER generation", () => {
-  let tmpdir: string;
-
-  beforeEach(() => {
-    tmpdir = fs.mkdtempSync(path.join(os.tmpdir(), "sizuku-e2e-mermaid-"));
-  });
-
-  afterEach(() => {
-    fs.rmSync(tmpdir, { recursive: true, force: true });
-  });
-
-  it("generates Mermaid ER with full erDiagram block for EC schema", async () => {
-    fs.writeFileSync(path.join(tmpdir, "schema.ts"), ecSchemaContent, "utf-8");
-    const output = path.join(tmpdir, "output", "ER.md");
-
-    const code = stripImports(ecSchemaContent);
-    const { sizukuMermaidER } = await import("../generator/mermaid-er/index.js");
-    const result = await sizukuMermaidER(code, output);
-
-    expect(result.ok).toBe(true);
-    expect(fs.existsSync(output)).toBe(true);
-
-    const content = fs.readFileSync(output, "utf-8");
-
-    // Mermaid structure
-    expect(content.startsWith("```mermaid\nerDiagram")).toBe(true);
-    expect(content.endsWith("```")).toBe(true);
-
-    // Relations (user -> order, order -> orderItem)
-    expect(content.includes('user ||--}| order : "(id) - (userId)"')).toBe(true);
-    expect(content.includes('order ||--}| orderItem : "(id) - (orderId)"')).toBe(true);
-
-    // Table definitions
-    expect(content.includes("user {")).toBe(true);
-    expect(content.includes("order {")).toBe(true);
-    expect(content.includes("orderItem {")).toBe(true);
-
-    // Field details
-    expect(content.includes("varchar id PK")).toBe(true);
-    expect(content.includes("varchar name")).toBe(true);
-    expect(content.includes("varchar userId FK")).toBe(true);
-    expect(content.includes("int orderId FK")).toBe(true);
-  });
-
-  it("generates Mermaid ER for single-table schema without relations", async () => {
-    fs.writeFileSync(path.join(tmpdir, "schema.ts"), singleTableSchema, "utf-8");
-    const output = path.join(tmpdir, "output", "ER.md");
-
-    const code = stripImports(singleTableSchema);
-    const { sizukuMermaidER } = await import("../generator/mermaid-er/index.js");
-    const result = await sizukuMermaidER(code, output);
-
-    expect(result.ok).toBe(true);
-    const content = fs.readFileSync(output, "utf-8");
-
-    expect(content.startsWith("```mermaid\nerDiagram")).toBe(true);
-    expect(content.endsWith("```")).toBe(true);
-    expect(content.includes("config {")).toBe(true);
-    expect(content.includes("text key PK")).toBe(true);
-    expect(content.includes("text value")).toBe(true);
-    // No relation lines (no ||-- pattern)
-    expect(content.includes("||--")).toBe(false);
-  });
-
-  it("generates Mermaid ER with field descriptions for auth schema", async () => {
-    fs.writeFileSync(path.join(tmpdir, "schema.ts"), pgAuthSchema, "utf-8");
-    const output = path.join(tmpdir, "output", "ER.md");
-
-    const code = stripImports(pgAuthSchema);
-    const { sizukuMermaidER } = await import("../generator/mermaid-er/index.js");
-    const result = await sizukuMermaidER(code, output);
-
-    expect(result.ok).toBe(true);
-    const content = fs.readFileSync(output, "utf-8");
-
-    expect(content.includes("account {")).toBe(true);
-    expect(content.includes('uuid id PK "Primary key"')).toBe(true);
-    expect(content.includes('varchar email "User email address"')).toBe(true);
-  });
-});
-
-// ============================================================================
-// E2E: PNG generation
-// ============================================================================
-
-describe("E2E: PNG generation", () => {
-  let tmpdir: string;
-
-  beforeEach(() => {
-    tmpdir = fs.mkdtempSync(path.join(os.tmpdir(), "sizuku-e2e-png-"));
-  });
-
-  afterEach(() => {
-    fs.rmSync(tmpdir, { recursive: true, force: true });
-  });
-
-  it("generates valid PNG binary for EC schema", async () => {
-    fs.writeFileSync(path.join(tmpdir, "schema.ts"), ecSchemaContent, "utf-8");
-    const output = path.join(tmpdir, "output", "er.png");
-
-    const code = stripImports(ecSchemaContent);
-    const { sizukuDbml } = await import("../generator/dbml/index.js");
-    const result = await sizukuDbml(code, output);
-
-    expect(result.ok).toBe(true);
-    expect(fs.existsSync(output)).toBe(true);
-
-    const buffer = fs.readFileSync(output);
-    // PNG magic bytes: 89 50 4E 47 (‰PNG)
-    expect(buffer[0]).toBe(0x89);
-    expect(buffer[1]).toBe(0x50);
-    expect(buffer[2]).toBe(0x4e);
-    expect(buffer[3]).toBe(0x47);
-    expect(buffer.length > 1000).toBe(true);
-  });
-
-  it("generates valid PNG for single-table schema", async () => {
-    fs.writeFileSync(path.join(tmpdir, "schema.ts"), singleTableSchema, "utf-8");
-    const output = path.join(tmpdir, "output", "config.png");
-
-    const code = stripImports(singleTableSchema);
-    const { sizukuDbml } = await import("../generator/dbml/index.js");
-    const result = await sizukuDbml(code, output);
-
-    expect(result.ok).toBe(true);
-    expect(fs.existsSync(output)).toBe(true);
-
-    const buffer = fs.readFileSync(output);
-    expect(buffer[0]).toBe(0x89);
-    expect(buffer[1]).toBe(0x50);
-    expect(buffer[2]).toBe(0x4e);
-    expect(buffer[3]).toBe(0x47);
-  });
-});
 
 // ============================================================================
 // E2E: Full pipeline via sizuku() with process.argv
