+++ packages/sizuku/src/cli/index.ts
@@ -1,10 +1,13 @@
+import {
+  sizukuArktype,
+  sizukuDbml,
+  sizukuEffect,
+  sizukuMermaidER,
+  sizukuValibot,
+  sizukuZod,
+} from "../core/index.js";
 import { readFileSync } from "../fsp/index.js";
-import { sizukuArktype } from "../generator/arktype/index.js";
-import { sizukuDbml } from "../generator/dbml/index.js";
-import { sizukuEffect } from "../generator/effect/index.js";
-import { sizukuMermaidER } from "../generator/mermaid-er/index.js";
-import { sizukuValibot } from "../generator/valibot/index.js";
-import { sizukuZod } from "../generator/zod/index.js";
+import { stripImports } from "../utils/index.js";
 
 const HELP_TEXT = `💧 sizuku - Drizzle ORM schema tools
 
@@ -23,15 +26,7 @@ Options:
   --no-with-relation                Do not generate relation schemas
   -h, --help                        Display this help message`;
 
-export function detectOutputType(output: string) {
-  if (output.endsWith(".dbml")) return "dbml" as const;
-  if (output.endsWith(".png")) return "png" as const;
-  if (output.endsWith(".md")) return "mermaid" as const;
-  if (output.endsWith(".ts")) return "typescript" as const;
-  return null;
-}
-
-export function parseFlags(argv: readonly string[]) {
+function parseFlags(argv: readonly string[]) {
   const ZOD_VERSIONS = ["v4", "mini", "@hono/zod-openapi"] as const;
   const zodVersionIndex = argv.findIndex(
     (a) => a === "--zod-version" || a.startsWith("--zod-version="),
@@ -53,104 +48,6 @@ export function parseFlags(argv: readonly string[]) {
     withComment: !argv.includes("--no-with-comment"),
     withRelation: !argv.includes("--no-with-relation"),
   };
-}
-
-export function stripImports(content: string) {
-  const lines = content.split("\n");
-  const codeStart = lines.findIndex(
-    (line) => !line.trim().startsWith("import") && line.trim() !== "",
-  );
-  return lines.slice(codeStart);
-}
-
-export function resolveSchemaLibrary(flags: {
-  readonly zod: boolean;
-  readonly valibot: boolean;
-  readonly arktype: boolean;
-  readonly effect: boolean;
-}) {
-  if (flags.zod) return { name: "zod", label: "Zod" } as const;
-  if (flags.valibot) return { name: "valibot", label: "Valibot" } as const;
-  if (flags.arktype) return { name: "arktype", label: "ArkType" } as const;
-  if (flags.effect) return { name: "effect", label: "Effect" } as const;
-  return null;
-}
-
-async function sizukuDirectDiagram(
-  input: string,
-  output: string,
-  outputType: "dbml" | "png" | "mermaid",
-) {
-  const contentResult = readFileSync(input);
-  if (!contentResult.ok) {
-    return { ok: false, error: `Failed to read input: ${contentResult.error}` } as const;
-  }
-
-  const code = stripImports(contentResult.value);
-
-  if (outputType === "dbml" || outputType === "png") {
-    const result = await sizukuDbml(code, output);
-    if (!result.ok) return result;
-    const label = outputType === "png" ? "PNG" : "DBML";
-    return { ok: true, value: `💧 Generated ${label} at: ${output}` } as const;
-  }
-
-  const result = await sizukuMermaidER(code, output);
-  if (!result.ok) return result;
-  return { ok: true, value: `💧 Generated Mermaid ER at: ${output}` } as const;
-}
-
-async function sizukuDirectSchema(
-  input: string,
-  output: string,
-  flags: {
-    readonly zod: boolean;
-    readonly valibot: boolean;
-    readonly arktype: boolean;
-    readonly effect: boolean;
-    readonly zodVersion: "v4" | "mini" | "@hono/zod-openapi" | undefined;
-    readonly exportTypes: boolean;
-    readonly withComment: boolean;
-    readonly withRelation: boolean;
-  },
-) {
-  const lib = resolveSchemaLibrary(flags);
-  if (!lib) {
-    return {
-      ok: false,
-      error: "Specify --zod, --valibot, --arktype, or --effect for .ts output",
-    } as const;
-  }
-
-  const contentResult = readFileSync(input);
-  if (!contentResult.ok) {
-    return { ok: false, error: `Failed to read input: ${contentResult.error}` } as const;
-  }
-
-  const code = stripImports(contentResult.value);
-
-  const result = await (() => {
-    if (lib.name === "zod") {
-      return sizukuZod(
-        code,
-        output,
-        flags.withComment,
-        flags.exportTypes,
-        flags.zodVersion,
-        flags.withRelation,
-      );
-    }
-    if (lib.name === "valibot") {
-      return sizukuValibot(code, output, flags.withComment, flags.exportTypes, flags.withRelation);
-    }
-    if (lib.name === "arktype") {
-      return sizukuArktype(code, output, flags.withComment, flags.exportTypes, flags.withRelation);
-    }
-    return sizukuEffect(code, output, flags.withComment, flags.exportTypes, flags.withRelation);
-  })();
-
-  if (!result.ok) return result;
-  return { ok: true, value: `💧 Generated ${lib.label} schema at: ${output}` } as const;
 }
 
 export async function sizuku() {
@@ -175,7 +72,15 @@ export async function sizuku() {
     return { ok: false, error: "Missing output file path after -o" } as const;
   }
 
-  const outputType = detectOutputType(output);
+  const outputType = output.endsWith(".dbml")
+    ? ("dbml" as const)
+    : output.endsWith(".png")
+      ? ("png" as const)
+      : output.endsWith(".md")
+        ? ("mermaid" as const)
+        : output.endsWith(".ts")
+          ? ("typescript" as const)
+          : null;
   if (!outputType) {
     return {
       ok: false,
@@ -183,9 +88,69 @@ export async function sizuku() {
     } as const;
   }
 
-  if (outputType === "typescript") {
-    return sizukuDirectSchema(input, output, parseFlags(argv));
+  const flags = outputType === "typescript" ? parseFlags(argv) : null;
+  const lib =
+    flags === null
+      ? null
+      : flags.zod
+        ? ({ name: "zod", label: "Zod" } as const)
+        : flags.valibot
+          ? ({ name: "valibot", label: "Valibot" } as const)
+          : flags.arktype
+            ? ({ name: "arktype", label: "ArkType" } as const)
+            : flags.effect
+              ? ({ name: "effect", label: "Effect" } as const)
+              : null;
+  if (outputType === "typescript" && !lib) {
+    return {
+      ok: false,
+      error: "Specify --zod, --valibot, --arktype, or --effect for .ts output",
+    } as const;
   }
 
-  return sizukuDirectDiagram(input, output, outputType);
+  const contentResult = readFileSync(input);
+  if (!contentResult.ok) {
+    return { ok: false, error: `Failed to read input: ${contentResult.error}` } as const;
+  }
+  const code = stripImports(contentResult.value);
+
+  if (outputType === "dbml" || outputType === "png") {
+    const result = await sizukuDbml(code, output);
+    if (!result.ok) return result;
+    const label = outputType === "png" ? "PNG" : "DBML";
+    return { ok: true, value: `💧 Generated ${label} at: ${output}` } as const;
+  }
+
+  if (outputType === "mermaid") {
+    const result = await sizukuMermaidER(code, output);
+    if (!result.ok) return result;
+    return { ok: true, value: `💧 Generated Mermaid ER at: ${output}` } as const;
+  }
+
+  // outputType === "typescript", lib and flags are non-null here
+  if (!lib || !flags) {
+    return { ok: false, error: "internal: missing lib/flags" } as const;
+  }
+
+  const result = await (() => {
+    if (lib.name === "zod") {
+      return sizukuZod(
+        code,
+        output,
+        flags.withComment,
+        flags.exportTypes,
+        flags.zodVersion,
+        flags.withRelation,
+      );
+    }
+    if (lib.name === "valibot") {
+      return sizukuValibot(code, output, flags.withComment, flags.exportTypes, flags.withRelation);
+    }
+    if (lib.name === "arktype") {
+      return sizukuArktype(code, output, flags.withComment, flags.exportTypes, flags.withRelation);
+    }
+    return sizukuEffect(code, output, flags.withComment, flags.exportTypes, flags.withRelation);
+  })();
+  if (!result.ok) return result;
+  return { ok: true, value: `💧 Generated ${lib.label} schema at: ${output}` } as const;
 }
