+++ packages/sizuku/src/core/valibot.ts
@@ -1,0 +1,12 @@
+import path from "node:path";
+import { emit, valibot } from "../generator/index.js";
+
+export function sizukuValibot(
+  code: string[],
+  output: string,
+  comment?: boolean,
+  type?: boolean,
+  relation?: boolean,
+) {
+  return emit(valibot(code, comment, type, relation), path.dirname(output), output);
+}
