--- packages/sizuku/src/generator/dbml/generator/index.ts
@@ -1,12 +1,0 @@
-export {
-  dbmlContent,
-  escapeNote,
-  formatConstraints,
-  makeColumn,
-  makeEnum,
-  makeRef,
-  makeRefName,
-  makeTable,
-  makeColumnConstraints,
-  mapDrizzleType,
-} from "./dbml-content.js";
