--- packages/sizuku/src/generator/mermaid-er/generator/index.ts
@@ -1,7 +1,0 @@
-export {
-  erContent,
-  erContentFromMergedSchema,
-  makeRelationLine,
-  removeDuplicateRelations,
-  simplifyType,
-} from "./er-content.js";
