--- packages/sizuku/src/generator/mermaid-er/index.ts
@@ -1,60 +1,0 @@
-import path from "node:path";
-import { mkdir, writeFile } from "../../fsp/index.js";
-import { erContent, erContentFromMergedSchema } from "./generator/index.js";
-import { extractRelationsFromSchema, parseTableInfo } from "./validator/index.js";
-
-export async function sizukuMermaidER(code: string[], output: string) {
-  const tables = parseTableInfo(code);
-  const relations = extractRelationsFromSchema(code);
-  const ERContent = erContent(relations, tables);
-
-  const mkdirResult = await mkdir(path.dirname(output));
-  if (!mkdirResult.ok) {
-    return { ok: false, error: mkdirResult.error } as const;
-  }
-  const writeFileResult = await writeFile(output, ERContent);
-  if (!writeFileResult.ok) {
-    return { ok: false, error: writeFileResult.error } as const;
-  }
-  return { ok: true, value: undefined } as const;
-}
-
-export async function sizukuMermaidERFromMerged(
-  schema: {
-    readonly tables: readonly {
-      readonly tableName: string;
-      readonly columns: readonly {
-        readonly name: string;
-        readonly sqlType: string;
-        readonly isPrimaryKey: boolean;
-        readonly isNotNull: boolean;
-        readonly annotations: readonly { readonly type: string; readonly value: string }[];
-      }[];
-      readonly foreignKeys: readonly {
-        readonly sourceColumns: readonly string[];
-        readonly foreignTable: string;
-        readonly foreignColumns: readonly string[];
-      }[];
-    }[];
-    readonly relations: readonly {
-      readonly type: "one" | "many";
-      readonly sourceTable: string;
-      readonly referencedTable: string;
-      readonly sourceColumns?: readonly string[];
-      readonly foreignColumns?: readonly string[];
-    }[];
-  },
-  output: string,
-) {
-  const ERContent = erContentFromMergedSchema(schema);
-
-  const mkdirResult = await mkdir(path.dirname(output));
-  if (!mkdirResult.ok) {
-    return { ok: false, error: mkdirResult.error } as const;
-  }
-  const writeFileResult = await writeFile(output, ERContent);
-  if (!writeFileResult.ok) {
-    return { ok: false, error: writeFileResult.error } as const;
-  }
-  return { ok: true, value: undefined } as const;
-}
