--- packages/sizuku/src/generator/zod/index.test.ts
@@ -1,609 +1,0 @@
-import fs from "node:fs";
-import fsp from "node:fs/promises";
-import { afterEach, describe, expect, it } from "vite-plus/test";
-import { relationZodCode, sizukuZod, zodCode } from "./index.js";
-
-// Test run
-// pnpm vitest run ./src/generator/zod/index.test.ts
-
-const TEST_CODE = [
-  "export const user = mysqlTable('user', {",
-  "  /// Primary key",
-  "  /// @z.uuid()",
-  "  /// @v.pipe(v.string(), v.uuid())",
-  "  id: varchar('id', { length: 36 }).primaryKey(),",
-  "  /// Display name",
-  "  /// @z.string().min(1).max(50)",
-  "  /// @v.pipe(v.string(), v.minLength(1), v.maxLength(50))",
-  "  name: varchar('name', { length: 50 }).notNull(),",
-  "})",
-  "",
-  "/// @relation user.id post.userId one-to-many",
-  "export const post = mysqlTable('post', {",
-  "  /// Primary key",
-  "  /// @z.uuid()",
-  "  /// @v.pipe(v.string(), v.uuid())",
-  "  id: varchar('id', { length: 36 }).primaryKey(),",
-  "  /// Article title",
-  "  /// @z.string().min(1).max(100)",
-  "  /// @v.pipe(v.string(), v.minLength(1), v.maxLength(100))",
-  "  title: varchar('title', { length: 100 }).notNull(),",
-  "  /// Body content (no length limit)",
-  "  /// @z.string()",
-  "  /// @v.string()",
-  "  content: varchar('content', { length: 65535 }).notNull(),",
-  "  /// Foreign key referencing User.id",
-  "  /// @z.uuid()",
-  "  /// @v.pipe(v.string(), v.uuid())",
-  "  userId: varchar('user_id', { length: 36 }).notNull(),",
-  "})",
-  "",
-  "export const userRelations = relations(user, ({ many }) => ({",
-  "  posts: many(post),",
-  "}))",
-  "",
-  "export const postRelations = relations(post, ({ one }) => ({",
-  "  user: one(user, {",
-  "    fields: [post.userId],",
-  "    references: [user.id],",
-  "  }),",
-  "}))",
-  "",
-];
-
-const TEST_CODE_WITH_OBJECT_TYPES = [
-  "/// @z.strictObject",
-  "/// @v.strictObject",
-  "export const user = mysqlTable('user', {",
-  "  /// Primary key",
-  "  /// @z.uuid()",
-  "  /// @v.pipe(v.string(), v.uuid())",
-  "  id: varchar('id', { length: 36 }).primaryKey(),",
-  "  /// Display name",
-  "  /// @z.string().min(1).max(50)",
-  "  /// @v.pipe(v.string(), v.minLength(1), v.maxLength(50))",
-  "  name: varchar('name', { length: 50 }).notNull(),",
-  "})",
-  "",
-  "/// @relation user.id post.userId one-to-many",
-  "/// @z.looseObject",
-  "/// @v.looseObject",
-  "export const post = mysqlTable('post', {",
-  "  /// Primary key",
-  "  /// @z.uuid()",
-  "  /// @v.pipe(v.string(), v.uuid())",
-  "  id: varchar('id', { length: 36 }).primaryKey(),",
-  "  /// Article title",
-  "  /// @z.string().min(1).max(100)",
-  "  /// @v.pipe(v.string(), v.minLength(1), v.maxLength(100))",
-  "  title: varchar('title', { length: 100 }).notNull(),",
-  "  /// Body content (no length limit)",
-  "  /// @z.string()",
-  "  /// @v.string()",
-  "  content: varchar('content', { length: 65535 }).notNull(),",
-  "  /// Foreign key referencing User.id",
-  "  /// @z.uuid()",
-  "  /// @v.pipe(v.string(), v.uuid())",
-  "  userId: varchar('user_id', { length: 36 }).notNull(),",
-  "})",
-  "",
-  "/// @z.strictObject",
-  "/// @v.strictObject",
-  "export const userRelations = relations(user, ({ many }) => ({",
-  "  posts: many(post),",
-  "}))",
-  "",
-  "/// @z.strictObject",
-  "/// @v.strictObject",
-  "export const postRelations = relations(post, ({ one }) => ({",
-  "  user: one(user, {",
-  "    fields: [post.userId],",
-  "    references: [user.id],",
-  "  }),",
-  "}))",
-  "",
-];
-
-// ============================================================================
-// Real-world use case tests - Pure function tests
-// ============================================================================
-
-describe("zodCode E-Commerce pattern", () => {
-  it.concurrent("generates Order schema with comments and type", () => {
-    const result = zodCode(
-      {
-        name: "order",
-        fields: [
-          { name: "id", definition: "z.uuid()", description: "Order ID" },
-          {
-            name: "status",
-            definition: "z.enum(['PENDING', 'CONFIRMED', 'SHIPPED', 'DELIVERED', 'CANCELLED'])",
-            description: "Order status",
-          },
-          {
-            name: "totalAmount",
-            definition: "z.number().int().nonnegative()",
-            description: "Total amount in cents",
-          },
-          {
-            name: "customerId",
-            definition: "z.uuid()",
-            description: "Foreign key referencing Customer.id",
-          },
-        ],
-      },
-      true,
-      true,
-    );
-    expect(result).toBe(`export const OrderSchema = z.object({/**
- * Order ID
- */
-id:z.uuid(),
-/**
- * Order status
- */
-status:z.enum(['PENDING', 'CONFIRMED', 'SHIPPED', 'DELIVERED', 'CANCELLED']),
-/**
- * Total amount in cents
- */
-totalAmount:z.number().int().nonnegative(),
-/**
- * Foreign key referencing Customer.id
- */
-customerId:z.uuid()})
-
-export type Order = z.infer<typeof OrderSchema>
-`);
-  });
-
-  it.concurrent("generates Order schema without comments", () => {
-    const result = zodCode(
-      {
-        name: "order",
-        fields: [
-          { name: "id", definition: "z.uuid()" },
-          {
-            name: "status",
-            definition: "z.enum(['PENDING', 'CONFIRMED', 'SHIPPED'])",
-          },
-          { name: "totalAmount", definition: "z.number().int()" },
-        ],
-      },
-      false,
-      false,
-    );
-    expect(result).toBe(`export const OrderSchema = z.object({id:z.uuid(),
-status:z.enum(['PENDING', 'CONFIRMED', 'SHIPPED']),
-totalAmount:z.number().int()})
-`);
-  });
-});
-
-describe("relationZodCode E-Commerce pattern", () => {
-  it.concurrent("generates Order relation with items and customer", () => {
-    const result = relationZodCode(
-      {
-        name: "orderRelations",
-        baseName: "order",
-        fields: [
-          { name: "items", definition: "z.array(OrderItemSchema)", description: undefined },
-          { name: "customer", definition: "CustomerSchema", description: undefined },
-        ],
-      },
-      true,
-    );
-    expect(result).toBe(`
-export const OrderRelationsSchema = z.object({...OrderSchema.shape,items:z.array(OrderItemSchema),customer:CustomerSchema})
-
-export type OrderRelations = z.infer<typeof OrderRelationsSchema>
-`);
-  });
-});
-
-describe("zodCode", () => {
-  it.concurrent("zodCode comment true type true", () => {
-    const result = zodCode(
-      {
-        name: "user",
-        fields: [
-          { name: "id", definition: "z.uuid()", description: "Primary key" },
-          {
-            name: "name",
-            definition: "z.string().min(1).max(50)",
-            description: "Display name",
-          },
-        ],
-      },
-      true,
-      true,
-    );
-
-    const expected = `export const UserSchema = z.object({/**
- * Primary key
- */
-id:z.uuid(),
-/**
- * Display name
- */
-name:z.string().min(1).max(50)})
-
-export type User = z.infer<typeof UserSchema>
-`;
-    expect(result).toBe(expected);
-  });
-
-  it.concurrent("zodCode comment false type false", () => {
-    const result = zodCode(
-      {
-        name: "user",
-        fields: [
-          { name: "id", definition: "z.uuid()", description: "Primary key" },
-          {
-            name: "name",
-            definition: "z.string().min(1).max(50)",
-            description: "Display name",
-          },
-        ],
-      },
-      false,
-      false,
-    );
-
-    const expected = `export const UserSchema = z.object({id:z.uuid(),
-name:z.string().min(1).max(50)})
-`;
-    expect(result).toBe(expected);
-  });
-
-  it.concurrent("zodCode strictObject", () => {
-    const result = zodCode(
-      {
-        name: "config",
-        fields: [{ name: "key", definition: "z.string()" }],
-        objectType: "strict",
-      },
-      false,
-      false,
-    );
-    expect(result).toBe(`export const ConfigSchema = z.strictObject({key:z.string()})\n`);
-  });
-
-  it.concurrent("zodCode looseObject", () => {
-    const result = zodCode(
-      {
-        name: "config",
-        fields: [{ name: "key", definition: "z.string()" }],
-        objectType: "loose",
-      },
-      false,
-      false,
-    );
-    expect(result).toBe(`export const ConfigSchema = z.looseObject({key:z.string()})\n`);
-  });
-
-  it.concurrent("zodCode strictObject with type", () => {
-    const result = zodCode(
-      {
-        name: "config",
-        fields: [{ name: "key", definition: "z.string()" }],
-        objectType: "strict",
-      },
-      false,
-      true,
-    );
-    expect(result).toBe(
-      `export const ConfigSchema = z.strictObject({key:z.string()})\n\nexport type Config = z.infer<typeof ConfigSchema>\n`,
-    );
-  });
-});
-
-describe("relationZodCode", () => {
-  it.concurrent("relationZodCode strict objectType strict", () => {
-    const result = relationZodCode(
-      {
-        name: "userRelations",
-        baseName: "user",
-        fields: [
-          {
-            name: "posts",
-            definition: "z.array(PostSchema)",
-            description: undefined,
-          },
-        ],
-        objectType: "strict",
-      },
-      true,
-    );
-    const expected = `
-export const UserRelationsSchema = z.strictObject({...UserSchema.shape,posts:z.array(PostSchema)})
-
-export type UserRelations = z.infer<typeof UserRelationsSchema>
-`;
-    expect(result).toBe(expected);
-  });
-  it.concurrent("relationZodCode objectType loose", () => {
-    const result = relationZodCode(
-      {
-        name: "userRelations",
-        baseName: "user",
-        fields: [
-          {
-            name: "posts",
-            definition: "z.array(PostSchema)",
-            description: undefined,
-          },
-        ],
-        objectType: "loose",
-      },
-      true,
-    );
-    const expected = `
-export const UserRelationsSchema = z.looseObject({...UserSchema.shape,posts:z.array(PostSchema)})
-
-export type UserRelations = z.infer<typeof UserRelationsSchema>
-`;
-    expect(result).toBe(expected);
-  });
-});
-
-// ============================================================================
-// Auth pattern - zodCode
-// ============================================================================
-
-describe("zodCode Auth pattern", () => {
-  it.concurrent("generates User schema with comments and type", () => {
-    const result = zodCode(
-      {
-        name: "user",
-        fields: [
-          { name: "id", definition: "z.uuid()", description: "Primary key" },
-          { name: "email", definition: "z.string().email()", description: "User email" },
-          {
-            name: "createdAt",
-            definition: "z.string().datetime()",
-            description: "Account creation timestamp",
-          },
-        ],
-      },
-      true,
-      true,
-    );
-    expect(result).toBe(`export const UserSchema = z.object({/**
- * Primary key
- */
-id:z.uuid(),
-/**
- * User email
- */
-email:z.string().email(),
-/**
- * Account creation timestamp
- */
-createdAt:z.string().datetime()})
-
-export type User = z.infer<typeof UserSchema>
-`);
-  });
-
-  it.concurrent("generates Session schema with comments and type", () => {
-    const result = zodCode(
-      {
-        name: "session",
-        fields: [
-          { name: "id", definition: "z.uuid()", description: "Session identifier" },
-          { name: "userId", definition: "z.uuid()", description: "Foreign key to User" },
-          {
-            name: "expiresAt",
-            definition: "z.string().datetime()",
-            description: "Session expiry",
-          },
-        ],
-      },
-      true,
-      true,
-    );
-    expect(result).toBe(`export const SessionSchema = z.object({/**
- * Session identifier
- */
-id:z.uuid(),
-/**
- * Foreign key to User
- */
-userId:z.uuid(),
-/**
- * Session expiry
- */
-expiresAt:z.string().datetime()})
-
-export type Session = z.infer<typeof SessionSchema>
-`);
-  });
-
-  it.concurrent("generates RBAC Role schema without comments", () => {
-    const result = zodCode(
-      {
-        name: "role",
-        fields: [
-          { name: "id", definition: "z.uuid()" },
-          { name: "name", definition: "z.string()" },
-        ],
-      },
-      false,
-      false,
-    );
-    expect(result).toBe(`export const RoleSchema = z.object({id:z.uuid(),
-name:z.string()})
-`);
-  });
-});
-
-describe("relationZodCode Auth pattern", () => {
-  it.concurrent("generates Session relation with user field", () => {
-    const result = relationZodCode(
-      {
-        name: "sessionRelations",
-        baseName: "session",
-        fields: [{ name: "user", definition: "UserSchema", description: undefined }],
-      },
-      true,
-    );
-    expect(result).toBe(`
-export const SessionRelationsSchema = z.object({...SessionSchema.shape,user:UserSchema})
-
-export type SessionRelations = z.infer<typeof SessionRelationsSchema>
-`);
-  });
-});
-
-describe("sizukuZod", () => {
-  afterEach(() => {
-    if (fs.existsSync("tmp-zod/zod-test.ts")) {
-      fs.unlinkSync("tmp-zod/zod-test.ts");
-    }
-    if (fs.existsSync("tmp-zod")) {
-      fs.rmdirSync("tmp-zod", { recursive: true });
-    }
-  });
-
-  it("sizukuZod", async () => {
-    await sizukuZod(TEST_CODE, "tmp-zod/zod-test.ts");
-    const result = await fsp.readFile("tmp-zod/zod-test.ts", "utf-8");
-    const expected = `import * as z from 'zod'
-
-export const UserSchema = z.object({ id: z.uuid(), name: z.string().min(1).max(50) })
-
-export const PostSchema = z.object({
-  id: z.uuid(),
-  title: z.string().min(1).max(100),
-  content: z.string(),
-  userId: z.uuid(),
-})
-`;
-    expect(result).toBe(expected);
-  });
-
-  it("sizukuZod comment true", async () => {
-    await sizukuZod(TEST_CODE, "tmp-zod/zod-test.ts", true);
-    const result = await fsp.readFile("tmp-zod/zod-test.ts", "utf-8");
-
-    const expected = `import * as z from 'zod'
-
-export const UserSchema = z.object({
-  /**
-   * Primary key
-   */
-  id: z.uuid(),
-  /**
-   * Display name
-   */
-  name: z.string().min(1).max(50),
-})
-
-export const PostSchema = z.object({
-  /**
-   * Primary key
-   */
-  id: z.uuid(),
-  /**
-   * Article title
-   */
-  title: z.string().min(1).max(100),
-  /**
-   * Body content (no length limit)
-   */
-  content: z.string(),
-  /**
-   * Foreign key referencing User.id
-   */
-  userId: z.uuid(),
-})
-`;
-    expect(result).toBe(expected);
-  });
-
-  it("sizukuZod type true", async () => {
-    await sizukuZod(TEST_CODE, "tmp-zod/zod-test.ts", false, true);
-    const result = await fsp.readFile("tmp-zod/zod-test.ts", "utf-8");
-    const expected = `import * as z from 'zod'
-
-export const UserSchema = z.object({ id: z.uuid(), name: z.string().min(1).max(50) })
-
-export type User = z.infer<typeof UserSchema>
-
-export const PostSchema = z.object({
-  id: z.uuid(),
-  title: z.string().min(1).max(100),
-  content: z.string(),
-  userId: z.uuid(),
-})
-
-export type Post = z.infer<typeof PostSchema>
-`;
-    expect(result).toBe(expected);
-  });
-
-  it("sizukuZod zod @hono/zod-openapi", async () => {
-    await sizukuZod(TEST_CODE, "tmp-zod/zod-test.ts", false, false, "@hono/zod-openapi");
-    const result = await fsp.readFile("tmp-zod/zod-test.ts", "utf-8");
-    const expected = `import { z } from '@hono/zod-openapi'
-
-export const UserSchema = z.object({ id: z.uuid(), name: z.string().min(1).max(50) })
-
-export const PostSchema = z.object({
-  id: z.uuid(),
-  title: z.string().min(1).max(100),
-  content: z.string(),
-  userId: z.uuid(),
-})
-`;
-    expect(result).toBe(expected);
-  });
-
-  it("sizukuZod with strictObject and looseObject", async () => {
-    await sizukuZod(TEST_CODE_WITH_OBJECT_TYPES, "tmp-zod/zod-test.ts");
-    const result = await fsp.readFile("tmp-zod/zod-test.ts", "utf-8");
-    const expected = `import * as z from 'zod'
-
-export const UserSchema = z.strictObject({ id: z.uuid(), name: z.string().min(1).max(50) })
-
-export const PostSchema = z.looseObject({
-  id: z.uuid(),
-  title: z.string().min(1).max(100),
-  content: z.string(),
-  userId: z.uuid(),
-})
-`;
-    expect(result).toBe(expected);
-  });
-
-  it("sizukuZod with strictObject and looseObject with relation", async () => {
-    await sizukuZod(
-      TEST_CODE_WITH_OBJECT_TYPES,
-      "tmp-zod/zod-test.ts",
-      false,
-      false,
-      undefined,
-      true,
-    );
-    const result = await fsp.readFile("tmp-zod/zod-test.ts", "utf-8");
-    const expected = `import * as z from 'zod'
-
-export const UserSchema = z.strictObject({ id: z.uuid(), name: z.string().min(1).max(50) })
-
-export const PostSchema = z.looseObject({
-  id: z.uuid(),
-  title: z.string().min(1).max(100),
-  content: z.string(),
-  userId: z.uuid(),
-})
-
-export const UserRelationsSchema = z.strictObject({
-  ...UserSchema.shape,
-  posts: z.array(PostSchema),
-})
-
-export const PostRelationsSchema = z.looseObject({ ...PostSchema.shape, user: UserSchema })
-`;
-    expect(result).toBe(expected);
-  });
-});
