+++ packages/sizuku/src/helper/index.ts
@@ -1,0 +1,2 @@
+export * from "./extract-schemas.js";
+export * from "./extract-tables.js";
