--- packages/sizuku/src/helper/runtime/index.ts
@@ -1,14 +1,0 @@
-export {
-  createEmptyCommentInfo,
-  extractCommentsFromSource,
-  parseAnnotations,
-} from "./comment-extractor.js";
-export { loadSchemaFromModule, loadSchemaFromPath } from "./schema-loader.js";
-export {
-  createMergedSchemaFromRuntime,
-  findColumn,
-  findTable,
-  getForeignKeysTo,
-  getTableRelations,
-  mergeSchemaWithComments,
-} from "./schema-merger.js";
