--- packages/sizuku/src/symbols.ts
@@ -1,12 +1,0 @@
-/**
- * Drizzle ORM internal symbols for accessing table metadata
- */
-
-export const AnyInlineForeignKeys = Symbol.for("drizzle:AnyInlineForeignKeys");
-export const PgInlineForeignKeys = Symbol.for("drizzle:PgInlineForeignKeys");
-export const MySqlInlineForeignKeys = Symbol.for("drizzle:MySqlInlineForeignKeys");
-export const SQLiteInlineForeignKeys = Symbol.for("drizzle:SQLiteInlineForeignKeys");
-export const TableName = Symbol.for("drizzle:Name");
-export const Schema = Symbol.for("drizzle:Schema");
-export const ExtraConfigBuilder = Symbol.for("drizzle:ExtraConfigBuilder");
-export const ExtraConfigColumns = Symbol.for("drizzle:ExtraConfigColumns");
