+++ packages/sizuku/src/utils/index.test.ts
@@ -1,23 +1,10 @@
 import { describe, expect, it } from "vite-plus/test";
-import { makeRelationLine, extractRelations, toRelationSymbol } from "../helper/extract-schemas.js";
 import {
   extractFieldComments,
   fieldDefinitions,
-  infer,
-  inferArktype,
-  inferEffect,
-  inferOutput,
   makeCapitalized,
-  makeCommentBlock,
-  makeValibotObject,
-  makeZodObject,
-  parseFieldComments,
-  parseRelationLine,
-  removeOptionalSuffix,
-  splitByTo,
-  splitByWhitespace,
   makeRelationFields,
-  resolveArktypeUndeclared,
+  parseFieldComments,
   resolveWrapperType,
 } from "./index";
 
@@ -142,229 +129,6 @@ export const postRelations = relations(post, ({ one }) => ({
     });
   });
 
-  describe("extractRelations", () => {
-    it.concurrent("extractRelations", () => {
-      const result = extractRelations(["/// @relation user.id post.userId one-to-many"]);
-      const expected = [
-        {
-          fromModel: "user",
-          fromField: "id",
-          toModel: "post",
-          toField: "userId",
-          type: "one-to-many",
-        },
-      ];
-      expect(result).toStrictEqual(expected);
-    });
-  });
-
-  describe("toRelationSymbol", () => {
-    it.concurrent("maps valid cardinalities", () => {
-      expect(toRelationSymbol("zero-one")).toBe("|o");
-      expect(toRelationSymbol("one")).toBe("||");
-      expect(toRelationSymbol("zero-many")).toBe("}o");
-      expect(toRelationSymbol("many")).toBe("}|");
-    });
-
-    it.concurrent("returns null for invalid cardinality", () => {
-      expect(toRelationSymbol("invalid")).toBeNull();
-      expect(toRelationSymbol("")).toBeNull();
-      expect(toRelationSymbol("one-to-many")).toBeNull();
-    });
-  });
-
-  describe("makeRelationLine", () => {
-    it.concurrent.each([
-      ["zero-one-to-zero-one", "|o--|o"],
-      ["zero-one-to-one", "|o--||"],
-      ["zero-one-to-zero-many", "|o--}o"],
-      ["zero-one-to-many", "|o--}|"],
-      ["zero-one-to-zero-one-optional", "|o..|o"],
-      ["zero-one-to-one-optional", "|o..||"],
-      ["zero-one-to-zero-many-optional", "|o..}o"],
-      ["zero-one-to-many-optional", "|o..}|"],
-      ["one-to-zero-one", "||--|o"],
-      ["one-to-one", "||--||"],
-      ["one-to-zero-many", "||--}o"],
-      ["one-to-many", "||--}|"],
-      ["one-to-zero-one-optional", "||..|o"],
-      ["one-to-one-optional", "||..||"],
-      ["one-to-zero-many-optional", "||..}o"],
-      ["one-to-many-optional", "||..}|"],
-      ["zero-many-to-zero-one", "}o--|o"],
-      ["zero-many-to-one", "}o--||"],
-      ["zero-many-to-zero-many", "}o--}o"],
-      ["zero-many-to-many", "}o--}|"],
-      ["zero-many-to-zero-one-optional", "}o..|o"],
-      ["zero-many-to-one-optional", "}o..||"],
-      ["zero-many-to-zero-many-optional", "}o..}o"],
-      ["zero-many-to-many-optional", "}o..}|"],
-      ["many-to-zero-one", "}|--|o"],
-      ["many-to-one", "}|--||"],
-      ["many-to-zero-many", "}|--}o"],
-      ["many-to-many", "}|--}|"],
-      ["many-to-zero-one-optional", "}|..|o"],
-      ["many-to-one-optional", "}|..||"],
-      ["many-to-zero-many-optional", "}|..}o"],
-      ["many-to-many-optional", "}|..}|"],
-      ["zero-many-to-zero-one-optional", "}o..|o"],
-      ["zero-many-to-one-optional", "}o..||"],
-      ["zero-many-to-zero-many-optional", "}o..}o"],
-      ["zero-many-to-many-optional", "}o..}|"],
-      ["many-to-zero-one-optional", "}|..|o"],
-      ["many-to-one-optional", "}|..||"],
-      ["many-to-zero-many-optional", "}|..}o"],
-      ["many-to-many-optional", "}|..}|"],
-      ["zero-one-to-zero-one-optional", "|o..|o"],
-      ["zero-one-to-one-optional", "|o..||"],
-    ])("makeRelationLine(%s) -> ok: %s", (input, expected) => {
-      const result = makeRelationLine(input);
-      expect(result).toStrictEqual({ ok: true, value: expected });
-    });
-
-    it.concurrent("returns error for invalid input format", () => {
-      const result = makeRelationLine("invalid-string");
-      expect(result).toStrictEqual({ ok: false, error: "Invalid input format: invalid-string" });
-    });
-
-    it.concurrent("returns error for empty string", () => {
-      const result = makeRelationLine("");
-      expect(result).toStrictEqual({ ok: false, error: "Invalid input format: " });
-    });
-
-    it.concurrent("returns error for invalid relationship type", () => {
-      const result = makeRelationLine("foo-to-bar");
-      expect(result).toStrictEqual({ ok: false, error: "Invalid relationship string: foo-to-bar" });
-    });
-
-    it.concurrent("returns error for partial valid input", () => {
-      const result = makeRelationLine("one-to-invalid");
-      expect(result).toStrictEqual({
-        ok: false,
-        error: "Invalid relationship string: one-to-invalid",
-      });
-    });
-  });
-
-  describe("makeZodObject", () => {
-    it.concurrent("default object", () => {
-      expect(makeZodObject("id:z.uuid()")).toBe("z.object({id:z.uuid()})");
-    });
-    it.concurrent("strictObject", () => {
-      expect(makeZodObject("id:z.uuid()", "strictObject")).toBe("z.strictObject({id:z.uuid()})");
-    });
-    it.concurrent("looseObject", () => {
-      expect(makeZodObject("id:z.uuid()", "looseObject")).toBe("z.looseObject({id:z.uuid()})");
-    });
-  });
-
-  describe("makeValibotObject", () => {
-    it.concurrent("default object", () => {
-      expect(makeValibotObject("id:v.string()")).toBe("v.object({id:v.string()})");
-    });
-    it.concurrent("strictObject", () => {
-      expect(makeValibotObject("id:v.string()", "strictObject")).toBe(
-        "v.strictObject({id:v.string()})",
-      );
-    });
-    it.concurrent("looseObject", () => {
-      expect(makeValibotObject("id:v.string()", "looseObject")).toBe(
-        "v.looseObject({id:v.string()})",
-      );
-    });
-  });
-
-  describe("parseRelationLine", () => {
-    it.concurrent("parses valid relation", () => {
-      expect(parseRelationLine("@relation user.id post.userId one-to-many")).toStrictEqual({
-        fromModel: "user",
-        fromField: "id",
-        toModel: "post",
-        toField: "userId",
-        type: "one-to-many",
-      });
-    });
-    it.concurrent("returns null for non-relation", () => {
-      expect(parseRelationLine("/// comment")).toBeNull();
-    });
-    it.concurrent("returns null for malformed relation", () => {
-      expect(parseRelationLine("@relation incomplete")).toBeNull();
-    });
-    it.concurrent("returns null for invalid field format", () => {
-      expect(parseRelationLine("@relation user post.userId one-to-many")).toBeNull();
-    });
-  });
-
-  describe("splitByTo", () => {
-    it.concurrent("splits by -to-", () => {
-      expect(splitByTo("one-to-many")).toStrictEqual(["one", "many"]);
-    });
-    it.concurrent("returns null when no -to-", () => {
-      expect(splitByTo("invalid")).toBeNull();
-    });
-  });
-
-  describe("removeOptionalSuffix", () => {
-    it.concurrent("removes -optional suffix", () => {
-      expect(removeOptionalSuffix("one-optional")).toBe("one");
-    });
-    it.concurrent("returns unchanged if no suffix", () => {
-      expect(removeOptionalSuffix("one")).toBe("one");
-    });
-  });
-
-  describe("splitByWhitespace", () => {
-    it.concurrent("splits by whitespace", () => {
-      expect(splitByWhitespace("  a  b  c  ")).toStrictEqual(["a", "b", "c"]);
-    });
-    it.concurrent("empty string", () => {
-      expect(splitByWhitespace("")).toStrictEqual([]);
-    });
-  });
-
-  describe("infer", () => {
-    it.concurrent("infer", () => {
-      expect(infer("User")).toBe("export type User = z.infer<typeof UserSchema>");
-    });
-    it.concurrent("capitalizes first letter", () => {
-      expect(infer("user")).toBe("export type User = z.infer<typeof UserSchema>");
-    });
-  });
-
-  describe("inferOutput", () => {
-    it.concurrent("inferOutput", () => {
-      expect(inferOutput("User")).toBe("export type User = v.InferOutput<typeof UserSchema>");
-    });
-  });
-
-  describe("inferArktype", () => {
-    it.concurrent("generates arktype infer", () => {
-      expect(inferArktype("User")).toBe("export type User = typeof UserSchema.infer");
-    });
-    it.concurrent("capitalizes", () => {
-      expect(inferArktype("post")).toBe("export type Post = typeof PostSchema.infer");
-    });
-  });
-
-  describe("inferEffect", () => {
-    it.concurrent("generates effect infer", () => {
-      expect(inferEffect("User")).toBe("export type UserEncoded = typeof UserSchema.Encoded");
-    });
-    it.concurrent("capitalizes", () => {
-      expect(inferEffect("post")).toBe("export type PostEncoded = typeof PostSchema.Encoded");
-    });
-  });
-
-  describe("makeCommentBlock", () => {
-    it("generates multi-line JSDoc", () => {
-      expect(makeCommentBlock("Primary key")).toBe("/**\n * Primary key\n */\n");
-    });
-
-    it("returns empty string for empty description", () => {
-      expect(makeCommentBlock("")).toBe("");
-    });
-  });
-
   describe("fieldDefinitions", () => {
     it.concurrent("fieldDefinitions comment true", () => {
       const result = fieldDefinitions(
@@ -418,21 +182,6 @@ name:z.string().min(1).max(50)`;
   // ============================================================================
   // Real-world use case tests
   // ============================================================================
-
-  describe("E-Commerce pattern - infer functions", () => {
-    it.concurrent("infer generates Order type", () => {
-      expect(infer("order")).toBe("export type Order = z.infer<typeof OrderSchema>");
-    });
-    it.concurrent("inferOutput generates Order type", () => {
-      expect(inferOutput("order")).toBe("export type Order = v.InferOutput<typeof OrderSchema>");
-    });
-    it.concurrent("inferArktype generates Order type", () => {
-      expect(inferArktype("order")).toBe("export type Order = typeof OrderSchema.infer");
-    });
-    it.concurrent("inferEffect generates OrderEncoded type", () => {
-      expect(inferEffect("order")).toBe("export type OrderEncoded = typeof OrderSchema.Encoded");
-    });
-  });
 
   describe("E-Commerce pattern - fieldDefinitions", () => {
     it.concurrent("generates Order fields with comments", () => {
@@ -523,79 +272,6 @@ note:z.string().optional()`);
       });
     });
   });
-
-  describe("E-Commerce pattern - parseRelationLine", () => {
-    it.concurrent("parses Order → OrderItem relation", () => {
-      expect(parseRelationLine("@relation order.id orderItem.orderId one-to-many")).toStrictEqual({
-        fromModel: "order",
-        fromField: "id",
-        toModel: "orderItem",
-        toField: "orderId",
-        type: "one-to-many",
-      });
-    });
-
-    it.concurrent("parses Customer → Order relation", () => {
-      expect(parseRelationLine("@relation customer.id order.customerId one-to-many")).toStrictEqual(
-        {
-          fromModel: "customer",
-          fromField: "id",
-          toModel: "order",
-          toField: "customerId",
-          type: "one-to-many",
-        },
-      );
-    });
-
-    it.concurrent("parses User → Profile one-to-one relation", () => {
-      expect(parseRelationLine("@relation user.id profile.userId one-to-one")).toStrictEqual({
-        fromModel: "user",
-        fromField: "id",
-        toModel: "profile",
-        toField: "userId",
-        type: "one-to-one",
-      });
-    });
-  });
-
-  describe("E-Commerce pattern - extractRelations", () => {
-    it.concurrent("extracts multiple EC relations", () => {
-      const result = extractRelations([
-        "/// @relation customer.id order.customerId one-to-many",
-        "export const order = mysqlTable('order', {",
-        "  /// @relation order.id orderItem.orderId one-to-many",
-        "  export const orderItem = mysqlTable('order_item', {",
-      ]);
-      expect(result).toStrictEqual([
-        {
-          fromModel: "customer",
-          fromField: "id",
-          toModel: "order",
-          toField: "customerId",
-          type: "one-to-many",
-        },
-        {
-          fromModel: "order",
-          fromField: "id",
-          toModel: "orderItem",
-          toField: "orderId",
-          type: "one-to-many",
-        },
-      ]);
-    });
-  });
-
-  describe("E-Commerce pattern - makeRelationLine", () => {
-    it.concurrent("one-to-many for Customer → Order", () => {
-      expect(makeRelationLine("one-to-many")).toStrictEqual({ ok: true, value: "||--}|" });
-    });
-    it.concurrent("one-to-one for User → Profile", () => {
-      expect(makeRelationLine("one-to-one")).toStrictEqual({ ok: true, value: "||--||" });
-    });
-    it.concurrent("many-to-many for Product → Tag", () => {
-      expect(makeRelationLine("many-to-many")).toStrictEqual({ ok: true, value: "}|--}|" });
-    });
-  });
 });
 
 describe("resolveWrapperType", () => {
@@ -633,20 +309,6 @@ describe("makeRelationFields", () => {
   });
 });
 
-describe("resolveArktypeUndeclared", () => {
-  it('returns "+":"reject", for strict', () => {
-    expect(resolveArktypeUndeclared("strict")).toBe('"+":"reject",');
-  });
-
-  it('returns "+":"ignore", for loose', () => {
-    expect(resolveArktypeUndeclared("loose")).toBe('"+":"ignore",');
-  });
-
-  it("returns empty string for undefined", () => {
-    expect(resolveArktypeUndeclared(undefined)).toBe("");
-  });
-});
-
 // ============================================================================
 // Edge case tests
 // ============================================================================
@@ -666,42 +328,6 @@ describe("makeCapitalized edge cases", () => {
 
   it.concurrent("numeric start", () => {
     expect(makeCapitalized("123abc")).toBe("123abc");
-  });
-});
-
-describe("splitByTo edge cases", () => {
-  it.concurrent("compound left side: zero-one-to-many-optional", () => {
-    expect(splitByTo("zero-one-to-many-optional")).toStrictEqual(["zero-one", "many-optional"]);
-  });
-
-  it.concurrent("-to- at start", () => {
-    expect(splitByTo("-to-test")).toStrictEqual(["", "test"]);
-  });
-
-  it.concurrent("-to- at end", () => {
-    expect(splitByTo("test-to-")).toStrictEqual(["test", ""]);
-  });
-
-  it.concurrent("multiple -to- occurrences returns first match", () => {
-    expect(splitByTo("a-to-b-to-c")).toStrictEqual(["a", "b-to-c"]);
-  });
-});
-
-describe("removeOptionalSuffix edge cases", () => {
-  it.concurrent("many-optional", () => {
-    expect(removeOptionalSuffix("many-optional")).toBe("many");
-  });
-
-  it.concurrent("zero-many-optional", () => {
-    expect(removeOptionalSuffix("zero-many-optional")).toBe("zero-many");
-  });
-
-  it.concurrent("optional without dash prefix", () => {
-    expect(removeOptionalSuffix("optional")).toBe("optional");
-  });
-
-  it.concurrent("-optional alone", () => {
-    expect(removeOptionalSuffix("-optional")).toBe("");
   });
 });
 
@@ -819,13 +445,5 @@ describe("SNS Pattern (User/Post/Comment/Like) - parseFieldComments", () => {
       description: "User who liked",
       objectType: undefined,
     });
-  });
-});
-
-describe("makeCommentBlock edge cases", () => {
-  it.concurrent("description with special characters", () => {
-    expect(makeCommentBlock("User's email (required)")).toBe(
-      "/**\n * User's email (required)\n */\n",
-    );
   });
 });
